Move isValidJavaClass.
