+++ jib-core/src/main/java/com/google/cloud/tools/jib/configuration/BuildConfiguration.java
@@ -21,12 +21,10 @@ import com.google.cloud.tools.jib.image.ImageReference;
 import com.google.cloud.tools.jib.image.json.BuildableManifestTemplate;
 import com.google.cloud.tools.jib.image.json.V22ManifestTemplate;
 import com.google.cloud.tools.jib.registry.credentials.RegistryCredentials;
-import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
 import java.util.List;
 import javax.annotation.Nullable;
-import javax.lang.model.SourceVersion;
 
 /** Immutable configuration options for the builder process. */
 public class BuildConfiguration {
@@ -188,19 +186,6 @@ public class BuildConfiguration {
           throw new IllegalStateException();
       }
     }
-  }
-
-  /**
-   * @param className the class name to check
-   * @return {@code true} if {@code className} is a valid Java class name; {@code false} otherwise
-   */
-  public static boolean isValidJavaClass(String className) {
-    for (String part : Splitter.on('.').split(className)) {
-      if (!SourceVersion.isIdentifier(part)) {
-        return false;
-      }
-    }
-    return true;
   }
 
   public static Builder builder(BuildLogger buildLogger) {
